LayoutTests/fast/forms/number/input-number-wheel-expected.txt

Test for wheel operations for <input type=number>

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Initial value is 0. We'll wheel up by 1:
PASS input.value is "1"
Wheel up by 100:
PASS input.value is "2"
Wheel down by 1:
PASS input.value is "1"
Wheel down by 256:
PASS input.value is "0"
Disabled input element:
PASS input.value is "0"
Read-only input element:
PASS input.value is "0"
No focus:
PASS input.value is "0"
PASS successfullyParsed is true

TEST COMPLETE
